\name{TreeGhats-package}
\alias{TreeGhats-package}
\alias{TreeGhats}
\docType{package}
\title{
\packageTitle{TreeGhats}
}
\description{
\packageDescription{TreeGhats}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{TreeGhats}
\packageIndices{TreeGhats}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{TreeGhats}

Maintainer: \packageMaintainer{TreeGhats}
}
\references{
Raevel, V., Ayyappan, N., Balachandran, N., Aravajy, S., Barathan, N., Ramesh, B. R. & Munoz, F. (2017) Taxonomic and ecological database of trees of Western Ghats - TreeGhatsData (ed Zenodo), https://zenodo.org/record/846291
}

\keyword{Western Ghats}
\keyword{tree}
\keyword{taxonomy}
\keyword{phylogeny}
\keyword{database}

