\name{TreeGhats.phylo}
\alias{TreeGhats.phylo}
\docType{data}
\title{
Global phylogeny of the TreeGhatsData database
}
\description{
The phylogenetic tree includes all the accepted taxon names of the TreeGhatsData database.
}
\usage{data("TreeGhats.phylo")}
\format{
  \code{TreeGhats.phylo} is an object of class \code{phylo}, which can be analyzed and plotted with tools available in \code{ape} package.
}
\details{
The phylogeny has been generated by applying \code{\link{create.phylo.all}} to the accepted names in \code{\link{TreeGhatsData}}.
It is based on the megaphylogeny of Qian and Jin (2016), and missing taxa are added with \code{\link{S.PhyloMaker}} with \code{scenarios = "S3"}.
}
\references{
Qian H. & Jin Y. (2016). An updated megaphylogeny of plants, a tool for generating plant phylogenies, and an analysis of phylogenetic community structure. Journal of Plant Ecology. 9(2):233-239.
}
\seealso{
\code{\link{create.phylo.all}}, \code{\link{S.PhyloMaker}}, \code{\link{TreeGhatsData}}
}
\examples{
data(TreeGhats.phylo)
}
\keyword{datasets}
