\name{TreeGhats-package}
\alias{TreeGhats-package}
\alias{TreeGhats}
\docType{package}
\title{
\packageTitle{TreeGhats}
}
\description{
\packageDescription{TreeGhats}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{TreeGhats}
\packageIndices{TreeGhats}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{TreeGhats}

Maintainer: \packageMaintainer{TreeGhats}
}
\references{
~~ Literature or other references for background information ~~
}
%%~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
%%~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
